requirements.txt
aiofiles
boto3==1.28.71
fastapi==0.104.0
fastapi_jwt==0.1.12
mypy_boto3_s3==1.28.55
passlib==1.7.4
pydantic==2.4.2
python-dotenv==1.0.0
SQLAlchemy==2.0.22
alembic~=1.12.1
uvicorn
psycopg2-binary
python-multipart
bcrypt
pydantic[email]
pandas~=2.1.2
requests~=2.31.0
dataclasses-json
sentry-sdk[fastapi]
httpx
pytest
pdfkit
jinja2
smsto